can login, logout, and update profile. needs msg showing status/error
