+++ package.json
@@ -13,9 +13,9 @@
     "debug": "~2.6.9",
     "express": "~4.16.0",
     "express-session": "^1.15.6",
-    "express-validator": "^5.0.3",
+    "express-validator": "^5.1.1",
     "hbs": "~4.0.1",
-    "http-errors": "~1.6.2",
+    "http-errors": "^1.6.3",
     "morgan": "~1.9.0",
     "nodemon": "^1.17.3",
     "npm": "^5.8.0",
