+++ public/stylesheets/style.css
@@ -27,6 +27,12 @@ a {
     font-size: 16px;
   }
 
+  textarea {
+    border: 1px solid black;
+    border-radius: 5px;
+    font-size: 16px;
+  }
+
   span {
     font-size: 20px;
   }
