+++ views/register.hbs
@@ -5,7 +5,7 @@
     <link rel='stylesheet' href='/stylesheets/style.css' />
   </head>
   <body>
-    <form action="/users/register" method="POST">
+    <form action="/register" method="POST">
       <label for="">First Name</label>
       <input class="inputBox" type="text" name="firstname"/>
       <label for="">Last Name</label>
